Refactor: Move frr-tools to cmd directory
Reorganize the project structure by moving frr-tools binaries to the
standard Go cmd/ directory layout:
- frr-tools/metrics -> cmd/metrics
- frr-tools/status -> cmd/status
- frr-tools/reloader -> cmd/reloader

All import paths and Dockerfile references have been updated
accordingly.

Reported-at: https://github.com/metallb/frr-k8s/issues/275
(cherry picked from commit f9baf2dc65396a1b0568d3d28e612e0fe09b5a63)
