+++ cmd/metrics/collector/bfd.go
@@ -5,7 +5,7 @@ package collector
 import (
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
-	"github.com/metallb/frr-k8s/frr-tools/metrics/vtysh"
+	"github.com/metallb/frr-k8s/cmd/metrics/vtysh"
 	"github.com/metallb/frr-k8s/internal/frr"
 	"github.com/prometheus/client_golang/prometheus"
 )
